+++ components/admin/ProjectEdit/GraphicEdit/GraphicEdit.test.js
@@ -1,7 +1,6 @@
 import { mount } from 'enzyme';
 import toJSON from 'enzyme-to-json';
-import wait from 'waait';
-import { MockedProvider } from '@apollo/react-testing';
+import { MockedProvider, wait } from '@apollo/react-testing';
 import { RouterContext } from 'next/dist/next-server/lib/router-context';
 
 import GraphicEdit from './GraphicEdit';
@@ -11,8 +10,10 @@ import {
   mocks,
   publishedMocks,
   props,
-  getFiles,
-  getSupportFiles,
+  getLocalAdditionalFiles,
+  getLocalEditableFiles,
+  localGraphicFiles,
+  supportFilesConfig,
   suppressActWarning,
 } from './testHelpers';
 
@@ -29,6 +30,16 @@ jest.mock( 'next/dynamic', () => () => 'GraphicProject' );
 jest.mock(
   'components/admin/ActionButtons/ActionButtons',
   () => function ActionButtons() { return ''; },
+);
+
+jest.mock(
+  'components/admin/ActionHeadline/ActionHeadline',
+  () => function ActionHeadline() { return ''; },
+);
+
+jest.mock(
+  'components/admin/ButtonPublish/ButtonPublish',
+  () => function ButtonPublish() { return ''; },
 );
 
 jest.mock(
@@ -212,28 +223,6 @@ describe( '<GraphicEdit />, when there is an existing DRAFT graphic project', ()
     await wait( 0 );
     wrapper.update();
     const supportFiles = wrapper.find( 'SupportFiles' );
-    const { data } = mocks[0].result;
-    const projectId = props.id;
-    const supportFilesConfig = [
-      {
-        headline: 'editable files',
-        helperText: 'Original files that may be edited and adapted as needed for reuse.',
-        files: getSupportFiles( {
-          data,
-          type: 'editable',
-          projectId,
-        } ),
-      },
-      {
-        headline: 'additional files',
-        helperText: 'Additional files may include transcript files, style guides, or other support files needed by internal staff in order to properly use these graphics.',
-        files: getSupportFiles( {
-          data,
-          type: 'additional',
-          projectId,
-        } ),
-      },
-    ];
 
     expect( supportFiles.exists() ).toEqual( true );
     expect( supportFiles.prop( 'projectId' ) ).toEqual( props.id );
@@ -246,17 +235,40 @@ describe( '<GraphicEdit />, when there is an existing DRAFT graphic project', ()
     await wait( 0 );
     wrapper.update();
     const graphicFiles = wrapper.find( 'GraphicFilesFormContainer' );
-    const { data } = mocks[0].result;
-    const projectId = props.id;
+    const { images } = mocks[0].result.data.graphicProject;
 
     expect( graphicFiles.exists() ).toEqual( true );
     expect( graphicFiles.prop( 'projectId' ) ).toEqual( props.id );
-    expect( graphicFiles.prop( 'files' ) )
-      .toEqual( getFiles( {
-        data,
-        type: 'images',
-        projectId,
-      } ) );
+    expect( graphicFiles.prop( 'files' ) ).toEqual( images );
+  } );
+
+  it( 'renders the bottom of page buttons', async () => {
+    await wait( 0 );
+    wrapper.update();
+    const actions = wrapper.find( '.actions' );
+    const headline = actions.find( '.headline' );
+    const modal = actions.find( 'Modal' );
+    const modalTrigger = mount( modal.prop( 'trigger' ) );
+    const btnPublish = actions.find( 'ButtonPublish' );
+
+    expect( actions.exists() ).toEqual( true );
+    expect( headline.props() ).toEqual( {
+      className: 'headline',
+      type: 'graphic project',
+      published: false,
+      updated: false,
+    } );
+    expect( modal.prop( 'content' ).props.children.type )
+      .toEqual( 'GraphicProject' );
+    expect( modalTrigger.name() ).toEqual( 'Button' );
+    expect( modalTrigger.prop( 'className' ) )
+      .toEqual( 'action-btn btn--preview' );
+    expect( modalTrigger.prop( 'content' ) ).toEqual( 'Preview' );
+    expect( modalTrigger.prop( 'primary' ) ).toEqual( true );
+    expect( modalTrigger.prop( 'disabled' ) ).toEqual( !props.id );
+    expect( btnPublish.prop( 'status' ) )
+      .toEqual( mocks[0].result.data.graphicProject.status );
+    expect( btnPublish.prop( 'disabled' ) ).toEqual( !props.id );
   } );
 } );
 
@@ -416,28 +428,6 @@ describe( '<GraphicEdit />, when there is an existing PUBLISHED graphic project'
     await wait( 0 );
     wrapper.update();
     const supportFiles = wrapper.find( 'SupportFiles' );
-    const { data } = mocks[0].result;
-    const projectId = props.id;
-    const supportFilesConfig = [
-      {
-        headline: 'editable files',
-        helperText: 'Original files that may be edited and adapted as needed for reuse.',
-        files: getSupportFiles( {
-          data,
-          type: 'editable',
-          projectId,
-        } ),
-      },
-      {
-        headline: 'additional files',
-        helperText: 'Additional files may include transcript files, style guides, or other support files needed by internal staff in order to properly use these graphics.',
-        files: getSupportFiles( {
-          data,
-          type: 'additional',
-          projectId,
-        } ),
-      },
-    ];
 
     expect( supportFiles.exists() ).toEqual( true );
     expect( supportFiles.prop( 'projectId' ) ).toEqual( props.id );
@@ -450,17 +440,40 @@ describe( '<GraphicEdit />, when there is an existing PUBLISHED graphic project'
     await wait( 0 );
     wrapper.update();
     const graphicFiles = wrapper.find( 'GraphicFilesFormContainer' );
-    const { data } = mocks[0].result;
-    const projectId = props.id;
+    const { images } = mocks[0].result.data.graphicProject;
 
     expect( graphicFiles.exists() ).toEqual( true );
     expect( graphicFiles.prop( 'projectId' ) ).toEqual( props.id );
-    expect( graphicFiles.prop( 'files' ) )
-      .toEqual( getFiles( {
-        data,
-        type: 'images',
-        projectId,
-      } ) );
+    expect( graphicFiles.prop( 'files' ) ).toEqual( images );
+  } );
+
+  it( 'renders the bottom of page buttons', async () => {
+    await wait( 0 );
+    wrapper.update();
+    const actions = wrapper.find( '.actions' );
+    const headline = actions.find( '.headline' );
+    const modal = actions.find( 'Modal' );
+    const modalTrigger = mount( modal.prop( 'trigger' ) );
+    const btnPublish = actions.find( 'ButtonPublish' );
+
+    expect( actions.exists() ).toEqual( true );
+    expect( headline.props() ).toEqual( {
+      className: 'headline',
+      type: 'graphic project',
+      published: true,
+      updated: false,
+    } );
+    expect( modal.prop( 'content' ).props.children.type )
+      .toEqual( 'GraphicProject' );
+    expect( modalTrigger.name() ).toEqual( 'Button' );
+    expect( modalTrigger.prop( 'className' ) )
+      .toEqual( 'action-btn btn--preview' );
+    expect( modalTrigger.prop( 'content' ) ).toEqual( 'Preview' );
+    expect( modalTrigger.prop( 'primary' ) ).toEqual( true );
+    expect( modalTrigger.prop( 'disabled' ) ).toEqual( !props.id );
+    expect( btnPublish.prop( 'status' ) )
+      .toEqual( publishedMocks[0].result.data.graphicProject.status );
+    expect( btnPublish.prop( 'disabled' ) ).toEqual( !props.id );
   } );
 } );
 
@@ -515,6 +528,12 @@ describe( '<GraphicEdit />, when there is no props.id and local files have been 
     wrapper.update();
 
     expect( graphicEdit().exists() ).toEqual( true );
+  } );
+
+  it( 'does not render the bottom of page buttons', () => {
+    const actions = wrapper.find( '.actions' );
+
+    expect( actions.exists() ).toEqual( false );
   } );
 
   it( 'renders ProjectHeader', async () => {
@@ -580,33 +599,21 @@ describe( '<GraphicEdit />, when there is no props.id and local files have been 
     await wait( 0 );
     wrapper.update();
     const supportFiles = wrapper.find( 'SupportFiles' );
-    const { data } = mocks[2].result;
-    const projectId = newProps.id;
-    const supportFilesConfig = [
+    const config = [
       {
-        headline: 'editable files',
-        helperText: 'Original files that may be edited and adapted as needed for reuse.',
-        files: getSupportFiles( {
-          data,
-          type: 'editable',
-          projectId,
-        } ),
+        ...supportFilesConfig[0],
+        files: getLocalEditableFiles(),
       },
       {
-        headline: 'additional files',
-        helperText: 'Additional files may include transcript files, style guides, or other support files needed by internal staff in order to properly use these graphics.',
-        files: getSupportFiles( {
-          data,
-          type: 'additional',
-          projectId,
-        } ),
+        ...supportFilesConfig[1],
+        files: getLocalAdditionalFiles(),
       },
     ];
 
     expect( supportFiles.exists() ).toEqual( true );
     expect( supportFiles.prop( 'projectId' ) ).toEqual( newProps.id );
-    supportFilesConfig.forEach( ( config, i ) => {
-      expect( supportFiles.prop( 'fileTypes' )[i] ).toEqual( config );
+    config.forEach( ( obj, i ) => {
+      expect( supportFiles.prop( 'fileTypes' )[i] ).toEqual( obj );
     } );
   } );
 
@@ -614,16 +621,9 @@ describe( '<GraphicEdit />, when there is no props.id and local files have been 
     await wait( 0 );
     wrapper.update();
     const graphicFiles = wrapper.find( 'GraphicFilesFormContainer' );
-    const { data } = mocks[2].result;
-    const projectId = newProps.id;
 
     expect( graphicFiles.exists() ).toEqual( true );
     expect( graphicFiles.prop( 'projectId' ) ).toEqual( newProps.id );
-    expect( graphicFiles.prop( 'files' ) )
-      .toEqual( getFiles( {
-        data,
-        type: 'images',
-        projectId,
-      } ) );
+    expect( graphicFiles.prop( 'files' ) ).toEqual( localGraphicFiles );
   } );
 } );